pkgs: support directory as package, add simple build support
Makes it so it uses the file directly for a file, and the folder's
default.nix for a folder. Previously a folder would end up with a
package for each file in it.

A file pkgs.nix was added to enable easy and fast building of packages.
Usable like `nix-build pkgs.nix -A test` for building the package
defined in test.nix