add server-sync module
A stupid simple script that allows you to have an upload folder, which
contains files to be uploaded to a server, and a data folder that
contains all the resulting data.

Works with multiple clients, such that
anybody can add to the upload folder and the others will get it in their
data folder.

Stuff needs to be deleted on the server.