Improved deploy workflow
With this change, every machine has its own repository that tracks the
config that's deployed on it. Before deployments to a machine, this
repository is pulled and merged. Only then the config is built. After it
is built, and the changes committed (automatically), it is pushed to the
machine's repository again, but only if the deploy succeeded.

This has the very nice advantage that I can deploy to the same machine
from various deployment machines, without undoing any changes of other
machines.

In order to prevent a nasty git history due to my frequent rebasing to
clean up the history, I'm now doing all deployments on local wip
branches (which never get rebased), while the master branch will be maintained
occasionally for the public repository